Add docblocks for the attribute bag implementations.
PHPDoc is fussy and doesn't want to allow nice inheritance of the
documentation, so there is a fair bit of copy/paste here.  The problems
seemed to come down to interface inheritance perhaps not being fully
supported.

I think the class documentation is the most important bit here, the rest
is just to help out IDEs and phpdoc generation to save user's a little
effort.

I'm not sure that the AttributeBagNamespaced not returning a fluent
interface for its set* methods is correct, but I documented it the way
it was implemented for now.